pyproject.toml
[tool.poetry]
name = "flask restful"
version = "0.1.0"
description = ""
authors = ["BrianAvendanoLuciano <brianavendanoluciano@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
flask = "^3.0.3"
flask-restful = "^0.3.10"
load-dotenv = "^0.1.0"
flask-sqlalchemy = "^3.1.1"
flask-migrate = "^4.0.7"
flask-marshmallow = "^1.2.1"
marshmallow-sqlalchemy = "^1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
